+++ google/cloud/monitoring_v3/types/group_service/requests.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.api import monitored_resource_pb2  # type: ignore
+from google.cloud.monitoring_v3.types import common
+from google.cloud.monitoring_v3.types import group as gm_group
+
+__manifest__ = ()
+
+
+__all__ = tuple(sorted(__manifest__))
