+++ google/cloud/monitoring_v3/types/metric_service/requests.py
@@ -21,27 +21,16 @@ from google.cloud.monitoring_v3.types import common
 from google.cloud.monitoring_v3.types import metric as gm_metric
 from google.rpc import status_pb2  # type: ignore
 
-
-__protobuf__ = proto.module(
-    package="google.monitoring.v3",
-    manifest={
-        "ListMonitoredResourceDescriptorsRequest",
-        "ListMonitoredResourceDescriptorsResponse",
-        "GetMonitoredResourceDescriptorRequest",
-        "ListMetricDescriptorsRequest",
-        "ListMetricDescriptorsResponse",
-        "GetMetricDescriptorRequest",
-        "CreateMetricDescriptorRequest",
-        "DeleteMetricDescriptorRequest",
-        "ListTimeSeriesRequest",
-        "ListTimeSeriesResponse",
-        "CreateTimeSeriesRequest",
-        "CreateTimeSeriesError",
-        "CreateTimeSeriesSummary",
-        "QueryTimeSeriesRequest",
-        "QueryTimeSeriesResponse",
-        "QueryErrorList",
-    },
+__manifest__ = (
+    "ListMonitoredResourceDescriptorsRequest",
+    "GetMonitoredResourceDescriptorRequest",
+    "ListMetricDescriptorsRequest",
+    "GetMetricDescriptorRequest",
+    "CreateMetricDescriptorRequest",
+    "DeleteMetricDescriptorRequest",
+    "ListTimeSeriesRequest",
+    "CreateTimeSeriesRequest",
+    "QueryTimeSeriesRequest",
 )
 
 
@@ -77,6 +66,7 @@ class ListMonitoredResourceDescriptorsRequest(proto.Message):
             method. Using this field causes the method to return
             additional results from the previous method call.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -96,35 +86,6 @@ class ListMonitoredResourceDescriptorsRequest(proto.Message):
     )
 
 
-class ListMonitoredResourceDescriptorsResponse(proto.Message):
-    r"""The ``ListMonitoredResourceDescriptors`` response.
-
-    Attributes:
-        resource_descriptors (Sequence[google.api.monitored_resource_pb2.MonitoredResourceDescriptor]):
-            The monitored resource descriptors that are available to
-            this project and that match ``filter``, if present.
-        next_page_token (str):
-            If there are more results than have been returned, then this
-            field is set to a non-empty value. To see the additional
-            results, use that value as ``page_token`` in the next call
-            to this method.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    resource_descriptors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=monitored_resource_pb2.MonitoredResourceDescriptor,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
 class GetMonitoredResourceDescriptorRequest(proto.Message):
     r"""The ``GetMonitoredResourceDescriptor`` request.
 
@@ -140,6 +101,7 @@ class GetMonitoredResourceDescriptorRequest(proto.Message):
             The ``[RESOURCE_TYPE]`` is a predefined type, such as
             ``cloudsql_database``.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -179,6 +141,7 @@ class ListMetricDescriptorsRequest(proto.Message):
             method. Using this field causes the method to return
             additional results from the previous method call.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -198,35 +161,6 @@ class ListMetricDescriptorsRequest(proto.Message):
     )
 
 
-class ListMetricDescriptorsResponse(proto.Message):
-    r"""The ``ListMetricDescriptors`` response.
-
-    Attributes:
-        metric_descriptors (Sequence[google.api.metric_pb2.MetricDescriptor]):
-            The metric descriptors that are available to the project and
-            that match the value of ``filter``, if present.
-        next_page_token (str):
-            If there are more results than have been returned, then this
-            field is set to a non-empty value. To see the additional
-            results, use that value as ``page_token`` in the next call
-            to this method.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    metric_descriptors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=metric_pb2.MetricDescriptor,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
 class GetMetricDescriptorRequest(proto.Message):
     r"""The ``GetMetricDescriptor`` request.
 
@@ -242,6 +176,7 @@ class GetMetricDescriptorRequest(proto.Message):
             An example value of ``[METRIC_ID]`` is
             ``"compute.googleapis.com/instance/disk/read_bytes_count"``.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -263,6 +198,7 @@ class CreateMetricDescriptorRequest(proto.Message):
             metric <https://cloud.google.com/monitoring/custom-metrics>`__
             descriptor.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -290,6 +226,7 @@ class DeleteMetricDescriptorRequest(proto.Message):
             An example of ``[METRIC_ID]`` is:
             ``"custom.googleapis.com/my_test_metric"``.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -359,9 +296,11 @@ class ListTimeSeriesRequest(proto.Message):
             method. Using this field causes the method to return
             additional results from the previous method call.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     class TimeSeriesView(proto.Enum):
         r"""Controls which fields are returned by ``ListTimeSeries``."""
+        __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
         FULL = 0
         HEADERS = 1
 
@@ -407,54 +346,6 @@ class ListTimeSeriesRequest(proto.Message):
     )
 
 
-class ListTimeSeriesResponse(proto.Message):
-    r"""The ``ListTimeSeries`` response.
-
-    Attributes:
-        time_series (Sequence[google.cloud.monitoring_v3.types.TimeSeries]):
-            One or more time series that match the filter
-            included in the request.
-        next_page_token (str):
-            If there are more results than have been returned, then this
-            field is set to a non-empty value. To see the additional
-            results, use that value as ``page_token`` in the next call
-            to this method.
-        execution_errors (Sequence[google.rpc.status_pb2.Status]):
-            Query execution errors that may have caused
-            the time series data returned to be incomplete.
-        unit (str):
-            The unit in which all ``time_series`` point values are
-            reported. ``unit`` follows the UCUM format for units as seen
-            in https://unitsofmeasure.org/ucum.html. If different
-            ``time_series`` have different units (for example, because
-            they come from different metric types, or a unit is absent),
-            then ``unit`` will be "{not_a_unit}".
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    time_series = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=gm_metric.TimeSeries,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-    execution_errors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=3,
-        message=status_pb2.Status,
-    )
-    unit = proto.Field(
-        proto.STRING,
-        number=5,
-    )
-
-
 class CreateTimeSeriesRequest(proto.Message):
     r"""The ``CreateTimeSeries`` request.
 
@@ -478,6 +369,7 @@ class CreateTimeSeriesRequest(proto.Message):
             The maximum number of ``TimeSeries`` objects per ``Create``
             request is 200.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -487,81 +379,6 @@ class CreateTimeSeriesRequest(proto.Message):
         proto.MESSAGE,
         number=2,
         message=gm_metric.TimeSeries,
-    )
-
-
-class CreateTimeSeriesError(proto.Message):
-    r"""DEPRECATED. Used to hold per-time-series error status.
-
-    Attributes:
-        time_series (google.cloud.monitoring_v3.types.TimeSeries):
-            DEPRECATED. Time series ID that resulted in the ``status``
-            error.
-        status (google.rpc.status_pb2.Status):
-            DEPRECATED. The status of the requested write operation for
-            ``time_series``.
-    """
-
-    time_series = proto.Field(
-        proto.MESSAGE,
-        number=1,
-        message=gm_metric.TimeSeries,
-    )
-    status = proto.Field(
-        proto.MESSAGE,
-        number=2,
-        message=status_pb2.Status,
-    )
-
-
-class CreateTimeSeriesSummary(proto.Message):
-    r"""Summary of the result of a failed request to write data to a
-    time series.
-
-    Attributes:
-        total_point_count (int):
-            The number of points in the request.
-        success_point_count (int):
-            The number of points that were successfully
-            written.
-        errors (Sequence[google.cloud.monitoring_v3.types.CreateTimeSeriesSummary.Error]):
-            The number of points that failed to be
-            written. Order is not guaranteed.
-    """
-
-    class Error(proto.Message):
-        r"""Detailed information about an error category.
-
-        Attributes:
-            status (google.rpc.status_pb2.Status):
-                The status of the requested write operation.
-            point_count (int):
-                The number of points that couldn't be written because of
-                ``status``.
-        """
-
-        status = proto.Field(
-            proto.MESSAGE,
-            number=1,
-            message=status_pb2.Status,
-        )
-        point_count = proto.Field(
-            proto.INT32,
-            number=2,
-        )
-
-    total_point_count = proto.Field(
-        proto.INT32,
-        number=1,
-    )
-    success_point_count = proto.Field(
-        proto.INT32,
-        number=2,
-    )
-    errors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=3,
-        message=Error,
     )
 
 
@@ -590,6 +407,7 @@ class QueryTimeSeriesRequest(proto.Message):
             method. Using this field causes the method to return
             additional results from the previous method call.
     """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
 
     name = proto.Field(
         proto.STRING,
@@ -609,73 +427,4 @@ class QueryTimeSeriesRequest(proto.Message):
     )
 
 
-class QueryTimeSeriesResponse(proto.Message):
-    r"""The ``QueryTimeSeries`` response.
-
-    Attributes:
-        time_series_descriptor (google.cloud.monitoring_v3.types.TimeSeriesDescriptor):
-            The descriptor for the time series data.
-        time_series_data (Sequence[google.cloud.monitoring_v3.types.TimeSeriesData]):
-            The time series data.
-        next_page_token (str):
-            If there are more results than have been returned, then this
-            field is set to a non-empty value. To see the additional
-            results, use that value as ``page_token`` in the next call
-            to this method.
-        partial_errors (Sequence[google.rpc.status_pb2.Status]):
-            Query execution errors that may have caused
-            the time series data returned to be incomplete.
-            The available data will be available in the
-            response.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    time_series_descriptor = proto.Field(
-        proto.MESSAGE,
-        number=8,
-        message=gm_metric.TimeSeriesDescriptor,
-    )
-    time_series_data = proto.RepeatedField(
-        proto.MESSAGE,
-        number=9,
-        message=gm_metric.TimeSeriesData,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=10,
-    )
-    partial_errors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=11,
-        message=status_pb2.Status,
-    )
-
-
-class QueryErrorList(proto.Message):
-    r"""This is an error detail intended to be used with INVALID_ARGUMENT
-    errors.
-
-    Attributes:
-        errors (Sequence[google.cloud.monitoring_v3.types.QueryError]):
-            Errors in parsing the time series query
-            language text. The number of errors in the
-            response may be limited.
-        error_summary (str):
-            A summary of all the errors.
-    """
-
-    errors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=gm_metric.QueryError,
-    )
-    error_summary = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-__all__ = tuple(sorted(__protobuf__.manifest))
+__all__ = tuple(sorted(__manifest__))
